e2e/virtual/workspaces: adapt to clusterworkspaces
